Implemented RedditHandler. Cache now uses normal string and not text_factory=bytes
