+++ dailyreleases/__main__.py
@@ -2,8 +2,8 @@
 The main terminal-based entry point. Invoke as `dailyreleases' or `python3 -m dailyreleases'.
 """
 
+from .main import Main
+
 if __name__ == "__main__":
-    from .main import Main
-#    from .stores import epic
+
     Main().run_main()
-    #epic.search("Gravity Circuit")
