+++ dailyreleases/cache.py
@@ -17,14 +17,13 @@ class Cache:
         connection = sqlite3.connect(CONFIG.DATA_DIR.joinpath("cache.sqlite"))
         # allow accessing rows by index and case-insensitively by name
         connection.row_factory = sqlite3.Row
-        # do not try to decode bytes as utf-8 strings
-        connection.text_factory = bytes
         self.connection = connection
         self.cache_time = timedelta(seconds=CONFIG.CONFIG["web"].getint(
             "cache_time"))
 
     def setup(self):
         logger.debug("Setting up cache.")
+        # TODO: table requests obsolete?
         self.connection.execute(
             """
             CREATE TABLE IF NOT EXISTS
@@ -46,11 +45,12 @@ class Cache:
         self.connection.commit()
 
     def clean(self, older_than_days=3):
+        # Removes PREs from Cache that are older than specified days
         cutoff_timestamp = (datetime.utcnow() - timedelta(
             days=older_than_days)).timestamp()
         self.connection.execute(
             """
-            DELETE FROM requests
+            DELETE FROM pres
             WHERE timestamp < :cutoff;
             """,
             {
@@ -81,7 +81,7 @@ class Cache:
             """
             SELECT dirname, nfo_link, timestamp
             FROM pres
-            WHERE dirname = :searched_dirname;
+            WHERE dirname = :dirname;
             """,
             {"dirname": dirname},
         ).fetchone()
